feat(classifier): V8 — short-prompt GPU coverage (+14 patterns, fix 3 FP patterns)
Closes the 150-char length-fallback dead zone. 95/95 routing accuracy.

New patterns: compare A-vs-B, walk-through/describe-full, plan-how-to,
draft/write document (README/ADR/post-mortem), IaC config generation
(nginx/k8s/terraform), SQL query generation, implement-in-language (C++),
build service/endpoint (multi-adj), recommended approaches, ML experiment
run, service wiring, end-to-end test creation, LYNX/autoencoder network targets.

Fixed false positives:
- how-to-configure: now requires 10+ chars after keyword
- how-do-I: narrowed to 'we' only (team questions, not individual run/check)
- why-does fail: added fail(ing) and broader debugging verb list

Harness: CL1(0.01ms) CL2 CL3 exp_DD all PASS. 66 total GPU patterns.

+++ src/cohezion/integrations/agentverse/bridge.py
@@ -149,7 +149,7 @@ class AgentVerseBridge:
         if not coherences:
             return 0.0
 
-        return sum(coherences) / len(coherences)
+        return float(sum(coherences) / len(coherences))
 
     def check_hiho_violations(self) -> list[CoherenceViolation]:
         """Check for HIHO band violations.
